@@ -6,7 +6,7 @@
 
     <groupId>org.monarchinitiative.svart</groupId>
     <artifactId>svart</artifactId>
-    <version>2.0.0-RC3</version>
+    <version>2.0.0-RC4</version>
     <description>a library to ease representation of variants and performing genomic manipulations</description>
 
     <name>svart</name>
@@ -51,7 +51,7 @@
     </distributionManagement>
 
     <properties>
-        <java.version>11</java.version>
+        <java.version>17</java.version>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
     </properties>
 
@@ -131,6 +131,9 @@
                 <groupId>org.apache.maven.plugins</groupId>
                 <artifactId>maven-surefire-plugin</artifactId>
                 <version>3.0.0-M5</version>
+                <configuration>
+                    <useModulePath>false</useModulePath>
+                </configuration>
             </plugin>
             <plugin>
                 <groupId>org.apache.maven.plugins</groupId>
