@@ -2,7 +2,6 @@ package org.monarchinitiative.svart;
 
 import org.monarchinitiative.svart.util.Seq;
 
-import java.nio.charset.StandardCharsets;
 import java.util.Objects;
 
 /**
@@ -13,26 +12,67 @@ public abstract class BaseGenomicVariant<T extends GenomicVariant> extends BaseG
     private final String id;
     private final String ref;
     private final String alt;
+    // breakend-specific identifiers
+    private final String mateId;
+    private final String eventId;
     // derived fields
     private final VariantType variantType;
     private final int changeLength;
 
-    protected BaseGenomicVariant(Contig contig, String id, Strand strand, Coordinates coordinates, String ref, String alt, int changeLength) {
+    protected BaseGenomicVariant(Contig contig, String id, Strand strand, Coordinates coordinates, String ref, String alt, int changeLength, String mateId, String eventId) {
         super(contig, strand, coordinates);
         this.id = IdCache.cacheId(id);
-        this.ref = AlleleCache.cacheAllele(VariantType.requireNonSymbolic(ref));
-        this.alt = AlleleCache.cacheAllele(VariantType.requireNonBreakend(alt));
+        this.ref = AlleleCache.cacheAllele(validateRefAllele(ref));
+        this.alt = AlleleCache.cacheAllele(validateAltAllele(alt));
+        this.mateId = Objects.requireNonNullElse(mateId, "");
+        this.eventId = Objects.requireNonNullElse(eventId, "");
         this.variantType = VariantType.parseType(ref, alt);
         this.changeLength = checkChangeLength(coordinates, changeLength, variantType);
     }
 
     protected BaseGenomicVariant(Builder<?> builder) {
-        this(builder.contig, builder.id, builder.strand, builder.coordinates, builder.ref, builder.alt, builder.changeLength);
+        this(builder.contig, builder.id, builder.strand, builder.coordinates, builder.ref, builder.alt, builder.changeLength, builder.mateId, builder.eventId);
+    }
+
+    private String validateRefAllele(String ref) {
+        return validateAllele(ref, true);
+    }
+
+    private String validateAltAllele(String alt) {
+        if (alt.length() == 1 && (alt.charAt(0) == '*' || alt.charAt(0) == '.')) {
+            // Sonar complains this is a blocker issue:
+            // "When a method is designed to return an invariant value, it may be poor design, but it shouldn’t
+            // adversely affect the outcome of your program. However, when it happens on all paths through the logic,
+            // it is surely a bug.
+            // This rule raises an issue when a method contains several return statements that all return the same value."
+            //
+            // In this case we're running multiple checks and if they pass we want to return the input early, otherwise
+            // an error will be thrown by the final test in validateAllele. The return value _should_ always be the input
+            // value, in this case!
+            return alt;
+        }
+        return VariantType.isSymbolic(alt) ? alt : validateAllele(alt, false);
+    }
+
+    private String validateAllele(String allele, boolean isRef) {
+        for (int i = 0; i < allele.length(); i++) {
+            if (!validAlleleChar(allele.charAt(i))) {
+                throw new IllegalArgumentException("Illegal " + (isRef ? "ref" : "alt") + " allele: " + allele);
+            }
+        }
+        return allele;
+    }
+
+    private boolean validAlleleChar(char c) {
+        return switch (c) {
+            case 'A', 'a', 'T', 't', 'C', 'c', 'G', 'g', 'N', 'n' -> true;
+            default -> false;
+        };
     }
 
     private int checkChangeLength(Coordinates coordinates, int changeLength, VariantType variantType) {
         if (variantType.baseType() == VariantType.DEL && changeLength >= 0) {
-            throw new IllegalArgumentException("Illegal DEL changeLength:" + changeLength + ". Should be < 0 given coordinates  " + changeCoordinates());
+            throw new IllegalArgumentException("Illegal DEL changeLength:" + changeLength + ". Should be < 0 given coordinates " + changeCoordinates());
         }
         if (variantType.baseType() == VariantType.INS && changeLength <= 0) {
             throw new IllegalArgumentException("Illegal INS changeLength:" + changeLength + ". Should be > 0 given coordinates " + changeCoordinates());
@@ -40,12 +80,14 @@ public abstract class BaseGenomicVariant<T extends GenomicVariant> extends BaseG
         if (variantType.baseType() == VariantType.DUP && changeLength <= 0) {
             throw new IllegalArgumentException("Illegal DUP!changeLength:" + changeLength + ". Should be > 0 given coordinates " + changeCoordinates());
         }
-        if (variantType.baseType() == VariantType.INV && changeLength != 0 && !isSymbolic()) {
-            // symbolic alleles may not be precise, so this can cause failures
-            throw new IllegalArgumentException("Illegal INV! changeLength:" + changeLength + ". Should be 0 given coordinates " + changeCoordinates());
-        }
-        if (ref.length() != coordinates.length() && !isSymbolic()) {
-            throw new IllegalArgumentException(coordinates + " (length " + coordinates.length() + ") ref=" + ref + ", alt=" + alt + " inconsistent with allele change length of " + changeLength);
+        if (!isSymbolic()) {
+            if (ref.length() != coordinates.length()) {
+                throw new IllegalArgumentException("Ref allele length of " + ref.length() + " inconsistent with " + coordinates + " (length " + coordinates.length() + ") ref=" + ref + ", alt=" + alt);
+            }
+            int expectedChangeLength = alt().length() - ref().length();
+            if (changeLength != expectedChangeLength) {
+                throw new IllegalArgumentException("Given changeLength of " + changeLength + " inconsistent with expected changeLength of " + expectedChangeLength + " for variant " + changeCoordinates());
+            }
         }
         return changeLength;
     }
@@ -55,7 +97,15 @@ public abstract class BaseGenomicVariant<T extends GenomicVariant> extends BaseG
     }
 
     protected static int calculateChangeLength(String ref, String alt) {
+        VariantType.requireNonSymbolic(alt);
         return alt.length() - ref.length();
+    }
+
+    protected static String requireLengthIfSymbolic(String alt) {
+        if (VariantType.isSymbolic(alt)) {
+            throw new IllegalArgumentException("Missing changeLength for symbolic alt allele " + alt);
+        }
+        return alt;
     }
 
     /**
@@ -73,7 +123,7 @@ public abstract class BaseGenomicVariant<T extends GenomicVariant> extends BaseG
         return start + ref.length() + Coordinates.endDelta(coordinateSystem);
     }
 
-    protected abstract T newVariantInstance(Contig contig, String id, Strand strand, Coordinates coordinates, String ref, String alt, int changeLength);
+    protected abstract T newVariantInstance(Contig contig, String id, Strand strand, Coordinates coordinates, String ref, String alt, int changeLength, String mateId, String eventId);
 
     @Override
     protected T newRegionInstance(Contig contig, Strand strand, Coordinates coordinates) {
@@ -88,12 +138,22 @@ public abstract class BaseGenomicVariant<T extends GenomicVariant> extends BaseG
         if (this.coordinateSystem() == coordinateSystem) {
             return (T) this;
         }
-        return newVariantInstance(contig(), id, strand(), coordinates().withCoordinateSystem(coordinateSystem), ref, alt, changeLength);
+        return newVariantInstance(contig(), id, strand(), coordinates().withCoordinateSystem(coordinateSystem), ref, alt, changeLength, mateId, eventId);
     }
 
     @Override
     public String id() {
         return id;
+    }
+
+    @Override
+    public String mateId() {
+        return mateId;
+    }
+
+    @Override
+    public String eventId() {
+        return eventId;
     }
 
     @Override
@@ -117,15 +177,20 @@ public abstract class BaseGenomicVariant<T extends GenomicVariant> extends BaseG
     }
 
     @Override
+    public boolean isBreakend() {
+        return variantType == VariantType.BND;
+    }
+
+    @Override
     @SuppressWarnings("unchecked")
     public T withStrand(Strand other) {
-        if (strand() == other) {
+        if (strand() == other || isBreakend()) {
             return (T) this;
         }
 
         String refRevComp = Seq.reverseComplement(ref);
         String altRevComp = isSymbolic() ? alt : Seq.reverseComplement(alt);
-        return newVariantInstance(contig(), id, other, coordinates().invert(contig()), refRevComp, altRevComp, changeLength);
+        return newVariantInstance(contig(), id, other, coordinates().invert(contig()), refRevComp, altRevComp, changeLength, mateId, eventId);
     }
 
     @Override
@@ -134,28 +199,37 @@ public abstract class BaseGenomicVariant<T extends GenomicVariant> extends BaseG
         if (o == null || getClass() != o.getClass()) return false;
         if (!super.equals(o)) return false;
         BaseGenomicVariant<?> that = (BaseGenomicVariant<?>) o;
-        return changeLength == that.changeLength && ref.equals(that.ref) && alt.equals(that.alt) && variantType == that.variantType;
+        return changeLength == that.changeLength && id.equals(that.id) && ref.equals(that.ref) && alt.equals(that.alt) && mateId.equals(that.mateId) && eventId.equals(that.eventId) && variantType == that.variantType;
     }
 
     @Override
     public int hashCode() {
-        return Objects.hash(super.hashCode(), ref, alt, variantType, changeLength);
+        return Objects.hash(super.hashCode(), id, ref, alt, mateId, eventId, variantType, changeLength);
     }
 
     @Override
     public String toString() {
-        return "BaseVariant{" +
-                "contig=" + contig().id() +
-                ", strand=" + strand() +
-                ", coordinateSystem=" + coordinateSystem() +
-                ", startPosition=" + start() +
-                ", endPosition=" + end() +
-                ", ref='" + ref + '\'' +
-                ", alt='" + alt + '\'' +
-                ", variantType=" + variantType +
-                ", length=" + length() +
-                ", changeLength=" + changeLength +
-                '}';
+        return "GenomicVariant{" +
+               "contig=" + contigId() +
+               ", id='" + id + '\'' +
+               ", strand=" + strand() +
+               ", " + formatCoordinates() +
+               ", ref='" + ref + '\'' +
+               ", alt='" + alt + '\'' +
+               ", variantType=" + variantType +
+               ", length=" + length() +
+               ", changeLength=" + changeLength +
+               mateIdStr() +
+               eventIdStr() +
+               '}';
+    }
+
+    private String mateIdStr() {
+        return mateId.isEmpty() ? "" : ", mateId=" + mateId;
+    }
+
+    private String eventIdStr() {
+        return eventId.isEmpty() ? "" : ", eventId=" + eventId;
     }
 
     public abstract static class Builder<T extends Builder<T>> extends BaseGenomicRegion.Builder<T> {
@@ -166,6 +240,10 @@ public abstract class BaseGenomicVariant<T extends GenomicVariant> extends BaseG
 
         protected int changeLength = 0;
 
+        // breakend-specific fields
+        protected String mateId = "";
+        protected String eventId = "";
+
         // n.b. this class does not offer the usual plethora of Builder options for each and every variable as they are
         // inherently linked to one-another and to allow this will more than likely ensure that objects are built in an
         // improper state. These methods are intended to allow subclasses to easily pass in the correct parameters so as
@@ -174,26 +252,61 @@ public abstract class BaseGenomicVariant<T extends GenomicVariant> extends BaseG
         @Override
         public T with(GenomicVariant genomicVariant) {
             Objects.requireNonNull(genomicVariant, "variant cannot be null");
-            return with(genomicVariant.contig(), genomicVariant.id(), genomicVariant.strand(), genomicVariant.coordinates(), genomicVariant.ref(), genomicVariant.alt(), genomicVariant.changeLength());
+            return with(genomicVariant.contig(),
+                    genomicVariant.id(),
+                    genomicVariant.strand(),
+                    genomicVariant.coordinates(),
+                    genomicVariant.ref(),
+                    genomicVariant.alt(),
+                    genomicVariant.changeLength(),
+                    genomicVariant.mateId(),
+                    genomicVariant.eventId());
         }
 
-        public T with(Contig contig, String id, Strand strand, CoordinateSystem coordinateSystem, int start, String ref, String alt) {
+        public T with(Contig contig, Strand strand, Coordinates coordinates, String ref, String alt) {
+            return with(contig, strand, coordinates, ref, requireLengthIfSymbolic(alt), calculateChangeLength(ref, alt));
+        }
+
+        public T with(Contig contig, Strand strand, CoordinateSystem coordinateSystem, int start, String ref, String alt) {
+            requireLengthIfSymbolic(alt);
             int end = calculateEnd(start, coordinateSystem, ref, alt);
             Coordinates coordinates = Coordinates.of(coordinateSystem, start, end);
-            return with(contig, id, strand, coordinates, ref, alt);
+            return with(contig, strand, coordinates, ref, alt, calculateChangeLength(ref, alt));
         }
 
-        public T with(Contig contig, String id, Strand strand, Coordinates coordinates, String ref, String alt) {
-            return with(contig, id, strand, coordinates, ref, alt, calculateChangeLength(ref, alt));
+        public T with(Contig contig, Strand strand, Coordinates coordinates, String ref, String alt, int changeLength) {
+            return with(contig, id, strand, coordinates, ref, alt, changeLength, mateId, eventId);
         }
 
-        public T with(Contig contig, String id, Strand strand, Coordinates coordinates, String ref, String alt, int changeLength) {
-            VariantType.requireNonBreakend(alt);
+        // keep this private to prevent excessive scoping of this method as that would basically nullify the point of having a builder.
+        private T with(Contig contig, String id, Strand strand, Coordinates coordinates, String ref, String alt, int changeLength, String mateId, String eventId) {
             super.with(contig, strand, coordinates);
             this.id = Objects.requireNonNull(id);
             this.ref = Objects.requireNonNull(ref);
             this.alt = Objects.requireNonNull(alt);
             this.changeLength = changeLength;
+            this.mateId = Objects.requireNonNullElse(mateId, this.mateId);
+            this.eventId = Objects.requireNonNullElse(eventId, this.eventId);
+            return self();
+        }
+
+        public T changeLength(int changeLength) {
+            this.changeLength = changeLength;
+            return self();
+        }
+
+        public T id(String id) {
+            this.id = Objects.requireNonNullElse(id, this.id);
+            return self();
+        }
+
+        public T mateId(String mateId) {
+            this.mateId = Objects.requireNonNullElse(mateId, this.mateId);
+            return self();
+        }
+
+        public T eventId(String eventId) {
+            this.eventId = Objects.requireNonNullElse(eventId, this.eventId);
             return self();
         }
 
@@ -246,26 +359,18 @@ public abstract class BaseGenomicVariant<T extends GenomicVariant> extends BaseG
         }
         
         private static String getCachedBase(String alt) {
-            byte base = alt.getBytes(StandardCharsets.UTF_8)[0];
-                switch (base) {
-                    case 'A':
-                        return A;
-                    case 'T':
-                        return T;
-                    case 'G':
-                        return G;
-                    case 'C':
-                        return C;
-                    case '.':
-                        return NO_CALL;
-                    case '*':
-                        return SPAN_DEL;
-                    case 'N':
-                        return N;
-                    default:
-                        return alt;
-                }
-            }
+            char base = alt.charAt(0);
+            return switch (base) {
+                case 'A' -> A;
+                case 'T' -> T;
+                case 'G' -> G;
+                case 'C' -> C;
+                case '.' -> NO_CALL;
+                case '*' -> SPAN_DEL;
+                case 'N' -> N;
+                default -> alt;
+            };
+        }
     }
 
     private static class IdCache {
