@@ -16,18 +16,7 @@ public abstract class BaseGenomicRegion<T extends GenomicRegion> implements Geno
         this.contig = Objects.requireNonNull(contig, "contig must not be null");
         this.strand = Objects.requireNonNull(strand, "strand must not be null");
         this.coordinates = Objects.requireNonNull(coordinates, "coordinates must not be null");
-        validateCoordinatesOnContig(coordinates, contig);
-    }
-
-    private void validateCoordinatesOnContig(Coordinates coordinates, Contig contig) {
-        CoordinateSystem coordinateSystem = coordinates.coordinateSystem();
-        int start = coordinates.start();
-        int end = coordinates.end();
-        if (coordinateSystem == CoordinateSystem.ONE_BASED && (start < 1 || end > contig.length())) {
-            throw new CoordinatesOutOfBoundsException("One-based coordinates " + contig.name() + ':' + start + '-' + end + " out of contig bounds [" + 1 + ',' + contig.length() + ']');
-        } else if (coordinateSystem == CoordinateSystem.ZERO_BASED && (start < 0 || end > contig.length())) {
-            throw new CoordinatesOutOfBoundsException("Zero-based coordinates " + contig.name() + ':' + start + '-' + end + " out of contig bounds [" + 0 + ',' + contig.length() + ')');
-        }
+        GenomicInterval.validateCoordinatesOnContig(coordinates, contig);
     }
 
     protected BaseGenomicRegion(Builder<?> builder) {
