@@ -16,12 +16,10 @@ import static org.monarchinitiative.svart.CoordinateSystem.ZERO_BASED;
  * @author Jules Jacobsen <j.jacobsen@qmul.ac.uk>
  * @author Daniel Danis <daniel.danis@jax.org>
  */
-public interface Coordinates extends CoordinateSystemed<Coordinates> {
-
-    Coordinates EMPTY = PreciseCoordinates.of(ZERO_BASED, 0, 0);
+public sealed interface Coordinates extends Convertible<Coordinates> permits PreciseCoordinates, ImpreciseCoordinates {
 
     default Coordinates empty() {
-        return EMPTY;
+        return PreciseCoordinates.EMPTY;
     }
 
     CoordinateSystem coordinateSystem();
@@ -100,7 +98,7 @@ public interface Coordinates extends CoordinateSystemed<Coordinates> {
         return coordinateSystem == ZERO_BASED ? 0 : ZERO_BASED.startDelta(coordinateSystem);
     }
 
-    Coordinates withPadding(int upstream, int downstream);
+    Coordinates extend(int upstream, int downstream);
 
     /**
      * Returns the length of a region, in bases
@@ -136,7 +134,8 @@ public interface Coordinates extends CoordinateSystemed<Coordinates> {
 
     /**
      * Determines whether two regions overlap, returning true if they do and false if they do not. Empty intervals are
-     * considered as overlapping if they are at the boundaries of the other interval. This method is transitive such
+     * NOT considered as overlapping if they are at the boundaries of the other interval. However, two empty intervals
+     * with the same start and end coordinates are considered as overlapping. This method is transitive such
      * that overlap(a, b) = overlap(b, a). The input {@link CoordinateSystem} are not required to match.
      *
      * @param aSystem {@link CoordinateSystem} for interval described by positions aStart and aEnd
@@ -148,12 +147,8 @@ public interface Coordinates extends CoordinateSystemed<Coordinates> {
      * @return true indicating intervals a and b overlap or false if they do not.
      */
     public static boolean overlap(CoordinateSystem aSystem, int aStart, int aEnd, CoordinateSystem bSystem, int bStart, int bEnd) {
-        // Check empty intervals abutting a region are included, this includes other empty intervals at the same position.
-        if (isEmpty(aSystem, aStart, aEnd)) {
-            return aContainsB(bSystem, bStart, bEnd, aSystem, aStart, aEnd);
-        }
-        if (isEmpty(bSystem, bStart, bEnd)) {
-            return aContainsB(aSystem, aStart, aEnd, bSystem, bStart, bEnd);
+        if (isEmpty(aSystem, aStart, aEnd) && isEmpty(bSystem, bStart, bEnd)) {
+            return zeroBasedStart(aSystem, aStart) == bEnd && zeroBasedStart(bSystem, bStart) == aEnd;
         }
         return zeroBasedStart(aSystem, aStart) < bEnd && zeroBasedStart(bSystem, bStart) < aEnd;
     }
