@@ -16,6 +16,22 @@ public interface GenomicVariant extends GenomicRegion {
     String id();
 
     /**
+     * For breakend variants. Returns the mateId of the mated breakend if set.
+     * @return The mateId or empty if not set.
+     */
+    default String mateId() {
+        return "";
+    }
+
+    /**
+     * For breakend variants. Returns the eventId of breakend if set.
+     * @return The eventId or empty if not set.
+     */
+    default String eventId() {
+        return "";
+    }
+
+    /**
      * @return String with the reference allele in the variant, without common
      * suffix or prefix to reference allele.
      */
@@ -47,7 +63,7 @@ public interface GenomicVariant extends GenomicRegion {
 
     @Override
     default GenomicVariant toOppositeStrand() {
-        return withStrand(strand().opposite());
+        return this.isBreakend() ? this : withStrand(strand().opposite());
     }
 
     default VariantType variantType() {
@@ -62,12 +78,12 @@ public interface GenomicVariant extends GenomicRegion {
         return VariantType.isBreakend(alt()) || variantType() == VariantType.BND;
     }
 
-    static Comparator<? super GenomicVariant> naturalOrder() {
+    static Comparator<GenomicVariant> naturalOrder() {
         return GenomicVariantNaturalOrderComparator.INSTANCE;
     }
 
     static int compare(GenomicVariant x, GenomicVariant y) {
-        int result = GenomicRegion.compare(x, y);
+        int result = GenomicInterval.compare(x, y);
         if (result == 0) {
             result = x.ref().compareTo(y.ref());
         }
@@ -80,28 +96,62 @@ public interface GenomicVariant extends GenomicRegion {
         return result;
     }
 
+    static GenomicVariant of(Contig contig, Strand strand, Coordinates coordinates, String ref, String alt) {
+        return of(contig, "", strand, coordinates, ref, alt);
+    }
+
     static GenomicVariant of(Contig contig, String id, Strand strand, Coordinates coordinates, String ref, String alt) {
-        VariantType.requireNonSymbolic(alt);
         return DefaultGenomicVariant.of(contig, id, strand, coordinates, ref, alt);
     }
 
     static GenomicVariant of(Contig contig, String id, Strand strand, CoordinateSystem coordinateSystem, int start, String ref, String alt) {
-        VariantType.requireNonSymbolic(alt);
         return DefaultGenomicVariant.of(contig, id, strand, coordinateSystem, start, ref, alt);
     }
 
+    static GenomicVariant of(Contig contig, Strand strand, CoordinateSystem coordinateSystem, int start, String ref, String alt) {
+        return DefaultGenomicVariant.of(contig, "", strand, coordinateSystem, start, ref, alt);
+    }
+
     static GenomicVariant of(Contig contig, String id, Strand strand, Coordinates coordinates, String ref, String alt, int changeLength) {
-        VariantType.requireSymbolic(alt);
         return DefaultGenomicVariant.of(contig, id, strand, coordinates, ref, alt, changeLength);
     }
 
+    static GenomicVariant of(Contig contig, Strand strand, Coordinates coordinates, String ref, String alt, int changeLength) {
+        return DefaultGenomicVariant.of(contig, "", strand, coordinates, ref, alt, changeLength);
+    }
+
+    static GenomicVariant of(Contig contig, Strand strand, CoordinateSystem coordinateSystem, int start, int end, String ref, String alt, int changeLength) {
+        Coordinates coordinates = Coordinates.of(coordinateSystem, start, end);
+        return DefaultGenomicVariant.of(contig, "", strand, coordinates, ref, alt, changeLength);
+    }
+
     static GenomicVariant of(Contig contig, String id, Strand strand, CoordinateSystem coordinateSystem, int start, int end, String ref, String alt, int changeLength) {
-        VariantType.requireSymbolic(alt);
         Coordinates coordinates = Coordinates.of(coordinateSystem, start, end);
         return DefaultGenomicVariant.of(contig, id, strand, coordinates, ref, alt, changeLength);
     }
 
+    static GenomicVariant of(Contig contig, String id, Strand strand, Coordinates coordinates, String ref, String alt, int changeLength, String mateId, String eventId) {
+        return DefaultGenomicVariant.of(contig, id, strand, coordinates, ref, alt, changeLength, mateId, eventId);
+    }
+
     static GenomicVariant of(String eventId, GenomicBreakend left, GenomicBreakend right, String ref, String alt) {
         return DefaultGenomicBreakendVariant.of(eventId, left, right, ref, alt);
+    }
+
+    static Builder builder() {
+        return new Builder();
+    }
+
+    class Builder extends BaseGenomicVariant.Builder<Builder> {
+
+        @Override
+        public BaseGenomicVariant<?> build() {
+            return new DefaultGenomicVariant(selfWithEndIfMissing());
+        }
+
+        @Override
+        protected GenomicVariant.Builder self() {
+            return this;
+        }
     }
 }
