@@ -2,36 +2,19 @@ package org.monarchinitiative.svart;
 
 import java.util.Objects;
 
-class PreciseCoordinates implements Coordinates, Comparable<Coordinates> {
+import static org.monarchinitiative.svart.CoordinateSystem.ZERO_BASED;
 
-    private final CoordinateSystem coordinateSystem;
-    private final int start;
-    private final int end;
+record PreciseCoordinates(CoordinateSystem coordinateSystem, int start, int end) implements Coordinates, Comparable<Coordinates> {
 
-    public PreciseCoordinates(CoordinateSystem coordinateSystem, int start, int end) {
-        this.coordinateSystem = Objects.requireNonNull(coordinateSystem);
-        this.start = start;
-        this.end = end;
-        Coordinates.validateCoordinates(this.coordinateSystem, this.start, this.end);
+    static final PreciseCoordinates EMPTY = new PreciseCoordinates(ZERO_BASED, 0, 0);
+
+    PreciseCoordinates {
+        Objects.requireNonNull(coordinateSystem);
+        Coordinates.validateCoordinates(coordinateSystem, start, end);
     }
 
     public static PreciseCoordinates of(CoordinateSystem coordinateSystem, int start, int end) {
         return new PreciseCoordinates(coordinateSystem, start, end);
-    }
-
-    @Override
-    public CoordinateSystem coordinateSystem() {
-        return coordinateSystem;
-    }
-
-    @Override
-    public int start() {
-        return start;
-    }
-
-    @Override
-    public int end() {
-        return end;
     }
 
     @Override
@@ -53,7 +36,7 @@ class PreciseCoordinates implements Coordinates, Comparable<Coordinates> {
     }
 
     @Override
-    public Coordinates withPadding(int upstream, int downstream) {
+    public Coordinates extend(int upstream, int downstream) {
         if (upstream == 0 && downstream == 0) {
             return this;
         }
@@ -66,24 +49,11 @@ class PreciseCoordinates implements Coordinates, Comparable<Coordinates> {
     }
 
     @Override
-    public boolean equals(Object o) {
-        if (this == o) return true;
-        if (o == null || getClass() != o.getClass()) return false;
-        PreciseCoordinates that = (PreciseCoordinates) o;
-        return start == that.start && end == that.end && coordinateSystem == that.coordinateSystem;
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hash(coordinateSystem, start, end);
-    }
-
-    @Override
     public String toString() {
         return "Coordinates{" +
-                "coordinateSystem=" + coordinateSystem +
-                ", start=" + start +
-                ", end=" + end +
-                '}';
+               "coordinateSystem=" + coordinateSystem +
+               ", start=" + start +
+               ", end=" + end +
+               '}';
     }
 }
