@@ -1,23 +1,13 @@
 package org.monarchinitiative.svart;
 
 /**
+ * Marker interface to flag a class as having a {@link Strand}. If the class requires the ability to be transposed onto
+ * the opposite strand, then it should implement the {@link Transposable} interface.
+ *
  * @author Jules Jacobsen <j.jacobsen@qmul.ac.uk>
  */
-public interface Stranded<T> {
+public interface Stranded {
 
     Strand strand();
 
-    T withStrand(Strand other);
-
-    default T toOppositeStrand() {
-        return withStrand(strand().opposite());
-    }
-
-    default T toPositiveStrand() {
-        return withStrand(Strand.POSITIVE);
-    }
-
-    default T toNegativeStrand() {
-        return withStrand(Strand.NEGATIVE);
-    }
 }
