@@ -1,8 +1,16 @@
 package org.monarchinitiative.svart;
 
+import org.monarchinitiative.svart.util.Seq;
+
 import java.util.Objects;
 
 /**
+ * VariantType follows the VCF types for structural variants or child terms of
+ * <a href="http://www.sequenceontology.org/browser/current_svn/term/SO:0001059">sequence_alteration</a> from the
+ * Sequence Ontology (SO), when not specified in VCF. These mostly match-up well, however there are a few terms
+ * for sequence variants (i.e. those with a known sequence) which do not have a VCF-specified type e.g. SNV, MNV, DELINS.
+ * In these cases the SO definition is followed.
+ *
  * @author Jules Jacobsen <j.jacobsen@qmul.ac.uk>
  */
 public enum VariantType {
@@ -16,13 +24,27 @@ public enum VariantType {
     //   A structural variation reference for medical and population genetics https://www.nature.com/articles/s41586-020-2287-8)
     // >150 (HTSJDK VariantContext.MAX_ALLELE_SIZE_FOR_NON_SV)
     // >=1000 (Jannovar VariantAnnotator)
+
+    /**
+     * SO:0001483 - SNVs are single nucleotide positions in genomic DNA at which different sequence alternatives exist.
+     */
     SNV,
-    MNV,    // a multi-nucleotide variation
-    // start, end, length
-    // pos
+
+    /**
+     * SO:0001013 - A multiple nucleotide polymorphism with alleles of common length > 1, for example AAA/TTT.
+     */
+    MNV,
+
+    /**
+     * SO:1000032 - A sequence alteration which included an insertion and a deletion, affecting 2 or more bases.
+     */
+    DELINS,
 
     SYMBOLIC,
     // VCF standard reserved values for structural variants
+    /**
+     * SO:0000159 - The point at which one or more contiguous nucleotides were excised.
+     */
     DEL,
     //DEL:ME
     DEL_ME(DEL),
@@ -32,6 +54,9 @@ public enum VariantType {
     DEL_ME_SVA(DEL, DEL_ME),
     DEL_ME_HERV(DEL, DEL_ME),
 
+    /**
+     * SO:0000667 - The sequence of one or more nucleotides added between two adjacent nucleotides in the sequence.
+     */
     INS,
     INS_ME(INS),
     INS_ME_ALU(INS, INS_ME),
@@ -90,21 +115,22 @@ public enum VariantType {
      * @return
      */
     public static VariantType parseType(String alt) {
+        Objects.requireNonNull(alt);
         if (alt.isEmpty()) {
             return UNKNOWN;
         }
-        String stripped = trimAngleBrackets(Objects.requireNonNull(alt));
+        String stripped = trimAngleBrackets(alt);
         switch (stripped) {
-            case "SNP":
-            case "SNV":
+            case "SNP", "SNV":
                 return SNV;
-            case "MNP":
-            case "MNV":
+            case "MNP",  "MNV":
                 return MNV;
             case "DEL":
                 return DEL;
             case "INS":
                 return INS;
+            case "DELINS":
+                return DELINS;
             case "DUP":
                 return DUP;
             case "INV":
@@ -198,8 +224,22 @@ public enum VariantType {
 
     /**
      * Returns the {@link VariantType} for the given ref/alt alleles. This method will determine whether the alleles
-     * indicate small sequence variations such as SNP/MNV or INS/DEL or if they indicate a large symbolic or breakend
-     * allele. It is required that the input alleles conform to the VCF specification.
+     * indicate small sequence variations such as SNP, MNV, INS, DEL or DELINS or if they indicate a large symbolic or
+     * breakend allele. It is required that the input alleles conform to the VCF specification, although this method will
+     * also handle fully-trimmed alleles where an INS or DEL contain an empty ref or alt allele. For small sequence variants,
+     * this method will assign a DELINS type to variants which can be described as a deletion, followed by an insertion
+     * event. For example A>GC is a deletion of the reference A followed by an insertion of a GC (a DELINS), whereas
+     * A>AGC is an INS of bases GC and AGC>A is a DEL of bases GC.
+     *
+     * <p>Note that, balanced substitutions such as TG>CA or TGA>CGC are typed as MNV following the Sequence Ontology
+     * definition <a href="http://www.sequenceontology.org/browser/current_svn/term/SO:0001013">SO:0001013</a>
+     * "A multiple nucleotide polymorphism with alleles of common length > 1, for example AAA/TTT".
+     * Similarly, DELINS also follows the <a href="http://www.sequenceontology.org/browser/current_svn/term/SO:1000032">SO:1000032</a>
+     * definition "A sequence alteration which included an insertion and a deletion, affecting 2 or more bases".
+     * The <a href="https://varnomen.hgvs.org/recommendations/DNA/variant/delins/">HGVS</a> definition of
+     * DELINS - "changes involving two or more consecutive nucleotides are described as deletion/insertion (delins)
+     * variants", would also include the MNV type.
+     * </p>
      *
      * @param ref reference allele string e.g. [ATGCN]+
      * @param alt alternate allele string e.g. a symbolic allele, or a breakend replacement string, or match the regular
@@ -207,22 +247,66 @@ public enum VariantType {
      * @return the {@link VariantType} calculated from the REF and ALT allele.
      */
     public static VariantType parseType(String ref, String alt) {
-        if (isSymbolic(ref, alt)) {
+        if (isSymbolic(alt)) {
             return parseType(alt);
         }
-        if (ref.length() == alt.length()) {
-            if (alt.length() == 1) {
-                return VariantType.SNV;
+        // SNV SO:0001483 - SNVs are single nucleotide positions in genomic DNA at which different sequence alternatives exist.
+        if (ref.length() == alt.length() && ref.length() == 1) {
+            return VariantType.SNV;
+        }
+        // MNV SO:0001013 - A multiple nucleotide polymorphism with alleles of common length > 1, for example AAA/TTT.
+        if (ref.length() == alt.length() && ref.length() > 1) {
+            // INV SO:1000036 - A continuous nucleotide sequence is inverted in the same position.
+            if (isInversion(ref, alt)) {
+                return VariantType.INV;
             }
             return VariantType.MNV;
         }
-        return ref.length() < alt.length() ? VariantType.INS : VariantType.DEL;
+        // DEL SO:0000159 - The point at which one or more contiguous nucleotides were excised.
+        // ATC>AT, AT>A or A> == DEL - handle untrimmed cases too
+        // check start and end of ref in case of flipping to opposite strands e.g. AG>A becomes CT>T
+        if (alt.length() < ref.length() && (ref.startsWith(alt) || ref.endsWith(alt))) {
+            return DEL;
+        }
+        // INS SO:0000667 - The sequence of one or more nucleotides added between two adjacent nucleotides in the sequence.
+        // >A or A>AT, AT>ATC == INS - handle untrimmed cases too
+        // check start and end of alt in case of flipping to opposite strands e.g. A>AG becomes T>CT
+        if (ref.length() < alt.length() && (alt.startsWith(ref) || alt.endsWith(ref))) {
+            return INS;
+        }
+        // DELINS SO:1000032 - A sequence alteration which included an insertion and a deletion, affecting 2 or more bases.
+        //  ATG>TC, AT>TCA  == DELINS (i.e. a deletion followed by an insertion)
+        return DELINS;
     }
 
-    public static boolean isSymbolic(String ref, String alt) {
-        // The VCF spec only mentions alt alleles as having symbolic characters, so check these first then check the ref
-        // just in case.
-        return isSymbolic(alt) || isSymbolic(ref);
+    /**
+     * Requires both ref and alt are <b>of equal length</b>
+     * @param ref
+     * @param alt
+     * @return
+     */
+    private static boolean isInversion(String ref, String alt) {
+        // check first and last bases are rev complement before checking the entire sequence
+        if (isReverseComplement(ref.charAt(0), alt.charAt(alt.length() -1)) && isReverseComplement(ref.charAt(ref.length() -1), alt.charAt(0))) {
+            if (ref.length() == 2) {
+                // Already checked the sequences are reverse complemented, although here AT -> AT, GC -> GC, so these are actually not variations!
+                // Other dinucleotide combinations are true inversions GA -> TC, CT -> AG
+                return true;
+            }
+            return Seq.reverseComplement(ref).equalsIgnoreCase(alt);
+        }
+        return false;
+    }
+
+    private static boolean isReverseComplement(char c, char c1) {
+        return switch (c) {
+            case 'A', 'a' -> c1 == 'T' || c1 == 't';
+            case 'T', 't' -> c1 == 'A' || c1 == 'a';
+            case 'C', 'c' -> c1 == 'G' || c1 == 'g';
+            case 'G', 'g' -> c1 == 'C' || c1 == 'c';
+            case 'N', 'n' -> c1 == 'N' || c1 == 'n';
+            default -> false;
+        };
     }
 
     public static boolean isSymbolic(String allele) {
@@ -234,7 +318,7 @@ public enum VariantType {
     }
 
     public static boolean isLargeSymbolic(String allele) {
-        return allele.length() > 1 && (allele.charAt(0) == '<' || allele.charAt(allele.length() - 1) == '>');
+        return allele.length() > 1 && (allele.charAt(0) == '<' && allele.charAt(allele.length() - 1) == '>');
     }
 
     public static boolean isSingleBreakend(String allele) {
@@ -247,31 +331,38 @@ public enum VariantType {
 
     public static String requireNonSymbolic(String alt) {
         if (alt == null || isSymbolic(alt)) {
-            throw new IllegalArgumentException("Illegal symbolic alt allele " + alt);
+            throw new IllegalArgumentException("Illegal symbolic alt allele '" + alt + "'");
         }
         if (alt.contains(",")) {
-            throw new IllegalArgumentException("Illegal multi-allelic alt allele " + alt);
+            throw new IllegalArgumentException("Illegal multi-allelic alt allele '" + alt + "'");
         }
         return alt;
     }
 
+    /**
+     * Checks the input allele string is a symbolic allele where symbolic includes structural variants of the form <DEL>,
+     * <INS> or uses breakend notation e.g. A[1:12345[ or A. or .A
+     * @param alt an alt allele string
+     * @return the input alt allele
+     * @throws IllegalArgumentException for an incorrect alt allele string
+     */
     public static String requireSymbolic(String alt) {
-        if (alt == null || !isLargeSymbolic(alt)) {
-            throw new IllegalArgumentException("Illegal non-symbolic or breakend alt allele " + alt);
+        if (alt == null || !isSymbolic(alt)) {
+            throw new IllegalArgumentException("Illegal non-symbolic alt allele '" + alt + "'");
         }
         return alt;
     }
 
     public static String requireBreakend(String alt) {
         if (alt == null || !isBreakend(alt)) {
-            throw new IllegalArgumentException("Illegal non-breakend alt allele " + alt);
+            throw new IllegalArgumentException("Illegal non-breakend alt allele '" + alt + "'");
         }
         return alt;
     }
 
     public static String requireNonBreakend(String alt) {
         if (alt == null || VariantType.isBreakend(alt)) {
-            throw new IllegalArgumentException("Illegal breakend allele " + alt);
+            throw new IllegalArgumentException("Illegal breakend allele '" + alt + "'");
         }
         return alt;
     }
