@@ -1,9 +1,6 @@
 package org.monarchinitiative.svart.interval;
 
-import org.monarchinitiative.svart.Contig;
-import org.monarchinitiative.svart.CoordinateSystem;
-import org.monarchinitiative.svart.GenomicRegion;
-import org.monarchinitiative.svart.Strand;
+import org.monarchinitiative.svart.*;
 
 import java.util.*;
 
@@ -20,42 +17,42 @@ import static java.util.stream.Collectors.toList;
  * @author Jules Jacobsen <j.jacobsen@qmul.ac.uk>
  * @since 2.0.0
  */
-public class GenomicRegionIndex<T extends GenomicRegion> {
+public class GenomicIntervalIndex<T extends GenomicInterval> {
     // RIGHT_OPEN, 0-start system
     private static final CoordinateSystem INDEX_COORDINATE_SYSTEM = IntervalTree.COORDINATE_SYSTEM;
     private static final Strand INDEX_STRAND = Strand.POSITIVE;
 
-    private static final GenomicRegionIndex<?> EMPTY = new GenomicRegionIndex<>(Map.of());
+    private static final GenomicIntervalIndex<?> EMPTY = new GenomicIntervalIndex<>(Map.of());
 
     private final Map<Contig, IntervalTree<T>> index;
 
-    private GenomicRegionIndex(Map<Contig, IntervalTree<T>> index) {
+    private GenomicIntervalIndex(Map<Contig, IntervalTree<T>> index) {
         this.index = index;
     }
 
     /**
-     * Static constructor for creating a {@link GenomicRegionIndex} from a collection of {@link GenomicRegionIndex}
+     * Static constructor for creating a {@link GenomicIntervalIndex} from a collection of {@link GenomicIntervalIndex}
      * objects of a given type.
      *
-     * @param genomicRegions The {@link GenomicRegion} objects to add to the index
+     * @param genomicIntervals The {@link GenomicRegion} objects to add to the index
      * @param <T> The type of {@link GenomicRegion} this index contains
-     * @return a {@link GenomicRegionIndex} containing the input {@link GenomicRegion} objects
+     * @return a {@link GenomicIntervalIndex} containing the input {@link GenomicRegion} objects
      */
-    public static <T extends GenomicRegion> GenomicRegionIndex<T> of(Collection<T> genomicRegions) {
-        Map<Contig, List<T>> regionIndex = genomicRegions.stream()
+    public static <T extends GenomicInterval> GenomicIntervalIndex<T> of(Collection<T> genomicIntervals) {
+        Map<Contig, List<T>> regionIndex = genomicIntervals.stream()
                 .distinct()
-                .sorted(GenomicRegion.naturalOrder())
+                .sorted(GenomicInterval.naturalOrder())
                 .collect(groupingBy(T::contig, toList()));
 
         Map<Contig, IntervalTree<T>> intervalTreeIndex = new HashMap<>();
         for (Map.Entry<Contig, List<T>> entry : regionIndex.entrySet()) {
-            intervalTreeIndex.put(entry.getKey(), new IntervalTree<>(entry.getValue(), new GenomicRegionNormaliser<>()));
+            intervalTreeIndex.put(entry.getKey(), new IntervalTree<>(entry.getValue(), new GenomicIntervalNormaliser<>()));
         }
 
-        return new GenomicRegionIndex<>(Map.copyOf(intervalTreeIndex));
+        return new GenomicIntervalIndex<>(Map.copyOf(intervalTreeIndex));
     }
 
-    private static class GenomicRegionNormaliser<T extends GenomicRegion> implements IntervalNormaliser<T> {
+    private static class GenomicIntervalNormaliser<T extends GenomicInterval> implements IntervalNormaliser<T> {
         @Override
         public int start(T x) {
             return x.startOnStrandWithCoordinateSystem(INDEX_STRAND, INDEX_COORDINATE_SYSTEM);
@@ -72,8 +69,8 @@ public class GenomicRegionIndex<T extends GenomicRegion> {
      * @return An empty index
      */
     @SuppressWarnings("unchecked")
-    public static <T extends GenomicRegion> GenomicRegionIndex<T> empty() {
-        return (GenomicRegionIndex<T>) EMPTY;
+    public static <T extends GenomicInterval> GenomicIntervalIndex<T> empty() {
+        return (GenomicIntervalIndex<T>) EMPTY;
     }
 
     /**
@@ -81,16 +78,16 @@ public class GenomicRegionIndex<T extends GenomicRegion> {
      * {@link GenomicRegion} on that {@link Contig}, a list of overlapping regions if present or the left and/or right
      * neighbouring {@link GenomicRegion} if there are no overlaps.
      *
-     * @param genomicRegion  The {@link GenomicRegion} of interest.
+     * @param genomicInterval  The {@link GenomicRegion} of interest.
      * @return A list of regions overlapping the given start and end positions.
      */
-    public IntervalOverlaps<T> regionsOverlapping(GenomicRegion genomicRegion) {
-        IntervalTree<T> intervalTree = index.get(genomicRegion.contig());
+    public IntervalOverlaps<T> regionsOverlapping(GenomicInterval genomicInterval) {
+        IntervalTree<T> intervalTree = index.get(genomicInterval.contig());
         if (intervalTree == null) {
             return IntervalOverlaps.empty();
         }
-        int begin = genomicRegion.startOnStrandWithCoordinateSystem(INDEX_STRAND, INDEX_COORDINATE_SYSTEM);
-        int end = genomicRegion.endOnStrandWithCoordinateSystem(INDEX_STRAND, INDEX_COORDINATE_SYSTEM);
+        int begin = genomicInterval.startOnStrandWithCoordinateSystem(INDEX_STRAND, INDEX_COORDINATE_SYSTEM);
+        int end = genomicInterval.endOnStrandWithCoordinateSystem(INDEX_STRAND, INDEX_COORDINATE_SYSTEM);
         return intervalTree.findOverlappingWithInterval(begin, end);
     }
 
@@ -106,7 +103,7 @@ public class GenomicRegionIndex<T extends GenomicRegion> {
     public boolean equals(Object o) {
         if (this == o) return true;
         if (o == null || getClass() != o.getClass()) return false;
-        GenomicRegionIndex<?> that = (GenomicRegionIndex<?>) o;
+        GenomicIntervalIndex<?> that = (GenomicIntervalIndex<?>) o;
         return Objects.equals(index, that.index);
     }
 
