@@ -1,47 +1,23 @@
 package org.monarchinitiative.svart.interval;
 
-import java.util.Objects;
-
 /**
  * Half-open interval for construction of an {@link IntervalTree}.
- *
+ * <p>
  * Taken from Jannovar, with minor alterations.
  *
+ * @param begin  start point of the interval (inclusive)
+ * @param end    end point of the interval (exclusive)
+ * @param value  the value stored for the Interval
+ * @param maxEnd the maximum of this nodes {@link #end} and both of its children's {@link #end}
  * @author <a href="mailto:manuel.holtgrewe@charite.de">Manuel Holtgrewe</a>
  * @since 2.0.0
  */
-public class Interval<T> implements Comparable<Interval<T>> {
-
-    /**
-     * start point of the interval (inclusive)
-     */
-    private final int begin;
-
-    /**
-     * end point of the interval (exclusive)
-     */
-    private final int end;
-
-    /**
-     * the value stored for the Interval
-     */
-    private final T value;
-
-    /**
-     * the maximum of this nodes {@link #end} and both of its children's {@link #end}
-     */
-    private final int maxEnd;
-
-    public Interval(int begin, int end, T value, int maxEnd) {
-        this.begin = begin;
-        this.end = end;
-        this.value = value;
-        this.maxEnd = maxEnd;
-    }
+public record IntervalTreeNode<T>(int begin, int end, T value, int maxEnd) implements Comparable<IntervalTreeNode<T>> {
 
     /**
      * @return start point of the interval (inclusive)
      */
+    @Override
     public int begin() {
         return begin;
     }
@@ -49,20 +25,15 @@ public class Interval<T> implements Comparable<Interval<T>> {
     /**
      * @return end point of the interval (exclusive)
      */
+    @Override
     public int end() {
         return end;
     }
 
     /**
-     * @return the value stored for the Interval
-     */
-    public T value() {
-        return value;
-    }
-
-    /**
      * @return the maximum of this nodes {@link #end} and both of it children's {@link #end}
      */
+    @Override
     public int maxEnd() {
         return maxEnd;
     }
@@ -102,20 +73,7 @@ public class Interval<T> implements Comparable<Interval<T>> {
         return begin < this.end && this.begin < end;
     }
 
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) return true;
-        if (o == null || getClass() != o.getClass()) return false;
-        Interval<?> interval = (Interval<?>) o;
-        return begin == interval.begin && end == interval.end && maxEnd == interval.maxEnd && value.equals(interval.value);
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hash(begin, end, value, maxEnd);
-    }
-
-    public int compareTo(Interval<T> o) {
+    public int compareTo(IntervalTreeNode<T> o) {
         final int result = begin - o.begin;
         if (result == 0)
             return end - o.end;
