@@ -6,7 +6,14 @@ import org.junit.jupiter.params.provider.CsvSource;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.equalTo;
 
-public class RegionTest {
+public class IntervalTest {
+
+    record BasicInterval(Coordinates coordinates) implements Interval {
+
+        static BasicInterval of(CoordinateSystem coordinateSystem, int start, int end) {
+            return new BasicInterval(Coordinates.of(coordinateSystem, start, end));
+        }
+    }
 
     @ParameterizedTest
     @CsvSource({
@@ -26,8 +33,8 @@ public class RegionTest {
     public void contains_region(CoordinateSystem coordinateSystem, int start, int end,
                                 CoordinateSystem queryCoordinateSystem, int queryStart, int queryEnd,
                                 boolean expected) {
-        TestRegion region = TestRegion.of(coordinateSystem, start, end);
-        TestRegion query = TestRegion.of(queryCoordinateSystem, queryStart, queryEnd);
+        Interval region = BasicInterval.of(coordinateSystem, start, end);
+        Interval query = BasicInterval.of(queryCoordinateSystem, queryStart, queryEnd);
 
         assertThat(region.contains(query), equalTo(expected));
     }
@@ -44,7 +51,7 @@ public class RegionTest {
             "ONE_BASED, 3, 3,   4, false",
     })
     public void contains_position(CoordinateSystem coordinateSystem, int start, int end, int pos, boolean expected) {
-        TestRegion region = TestRegion.of(coordinateSystem, start, end);
+        Interval region = BasicInterval.of(coordinateSystem, start, end);
         assertThat(region.contains(pos), equalTo(expected));
     }
 
@@ -59,8 +66,8 @@ public class RegionTest {
     public void overlapsWith(CoordinateSystem coordinateSystem, int start, int end,
                              CoordinateSystem queryCoordinateSystem, int queryStart, int queryEnd,
                              boolean expected) {
-        TestRegion region = TestRegion.of(coordinateSystem, start, end);
-        TestRegion query = TestRegion.of(queryCoordinateSystem, queryStart, queryEnd);
+        Interval region = BasicInterval.of(coordinateSystem, start, end);
+        Interval query = BasicInterval.of(queryCoordinateSystem, queryStart, queryEnd);
 
         assertThat(region.overlapsWith(query), equalTo(expected));
     }
@@ -79,24 +86,24 @@ public class RegionTest {
     public void distanceTo(CoordinateSystem coordinateSystem, int start, int end,
                            CoordinateSystem queryCoordinateSystem, int queryStart, int queryEnd,
                            int expected) {
-        TestRegion region = TestRegion.of(coordinateSystem, start, end);
-        TestRegion query = TestRegion.of(queryCoordinateSystem, queryStart, queryEnd);
+        Interval region = BasicInterval.of(coordinateSystem, start, end);
+        Interval query = BasicInterval.of(queryCoordinateSystem, queryStart, queryEnd);
 
         assertThat(region.distanceTo(query), equalTo(expected));
     }
 
     @ParameterizedTest
     @CsvSource({
-            "ZERO_BASED,    2, 2,   0",
-            "ONE_BASED, 2, 1,   0",
+            "ZERO_BASED, 2, 2,   0",
+            "ONE_BASED,  2, 1,   0",
 
-            "ZERO_BASED,    1, 2,   1",
-            "ONE_BASED, 2, 2,   1",
+            "ZERO_BASED, 1, 2,   1",
+            "ONE_BASED,  2, 2,   1",
 
-            "ONE_BASED, 2, 3,   2",
-            "ONE_BASED, 2, 4,   3",
+            "ONE_BASED,  2, 3,   2",
+            "ONE_BASED,  2, 4,   3",
     })
     public void length(CoordinateSystem coordinateSystem, int start, int end, int expected) {
-        assertThat(TestRegion.of(coordinateSystem,start, end).length(), equalTo(expected));
+        assertThat(BasicInterval.of(coordinateSystem,start, end).length(), equalTo(expected));
     }
 }
