@@ -9,13 +9,13 @@ public final class TestGenomicVariant extends BaseGenomicVariant<TestGenomicVari
         super(builder);
     }
 
-    public TestGenomicVariant(Contig contig, String id, Strand strand, Coordinates coordinates, String ref, String alt, int changeLength) {
-        super(contig, id, strand, coordinates, ref, alt, changeLength);
+    public TestGenomicVariant(Contig contig, String id, Strand strand, Coordinates coordinates, String ref, String alt, int changeLength, String mateId, String eventId) {
+        super(contig, id, strand, coordinates, ref, alt, changeLength, mateId, eventId);
     }
 
     @Override
-    protected TestGenomicVariant newVariantInstance(Contig contig, String id, Strand strand, Coordinates coordinates, String ref, String alt, int changeLength) {
-        return new TestGenomicVariant(contig, id, strand, coordinates, ref, alt, changeLength);
+    protected TestGenomicVariant newVariantInstance(Contig contig, String id, Strand strand, Coordinates coordinates, String ref, String alt, int changeLength, String mateId, String eventId) {
+        return new TestGenomicVariant(contig, id, strand, coordinates, ref, alt, changeLength, mateId, eventId);
     }
 
     public static Builder builder() {
