@@ -5,12 +5,12 @@ import org.junit.jupiter.api.Test;
 import org.monarchinitiative.svart.*;
 import org.monarchinitiative.svart.assembly.GenomicAssembly;
 import org.monarchinitiative.svart.impl.DefaultGenomicBreakendVariant;
-import org.monarchinitiative.svart.impl.DefaultGenomicVariant;
 
 import java.nio.file.Path;
 
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.equalTo;
+import static org.junit.jupiter.api.Assertions.assertSame;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 
 public class VcfConverterTest {
@@ -34,7 +34,7 @@ public class VcfConverterTest {
 
         @Test
         public void convertWithBuilder() {
-            DefaultGenomicVariant.Builder builder = instance.convert(DefaultGenomicVariant.builder(), instance.parseContig("chr1"), "rs123456", 12345, "C", "T");
+            GenomicVariant.Builder builder = instance.convert(GenomicVariant.builder(), instance.parseContig("chr1"), "rs123456", 12345, "C", "T");
             GenomicVariant variant = builder.build();
             assertThat(variant, equalTo(GenomicVariant.of(chr1, "rs123456", Strand.POSITIVE, CoordinateSystem.ONE_BASED, 12345, "C", "T")));
         }
@@ -52,19 +52,19 @@ public class VcfConverterTest {
             // CHR	POS	ID	REF	ALT
             // chr1	12345	rs123456	CCC	TC, TCC	6	PASS	.
             Exception exception = assertThrows(IllegalArgumentException.class, () -> instance.convert(instance.parseContig("chr1"), "rs123456", 12345, "CCC", "TC, TCC"));
-            assertThat(exception.getMessage(), equalTo("Illegal multi-allelic alt allele TC, TCC"));
+            assertThat(exception.getMessage(), equalTo("Illegal multi-allelic alt allele 'TC, TCC'"));
         }
 
         @Test
         public void throwsExceptionWithSymbolicAllele() {
             Exception exception = assertThrows(IllegalArgumentException.class, () -> instance.convert(instance.parseContig("chr1"), "rs123456", 12345, "C", "<DEL>"));
-            assertThat(exception.getMessage(), equalTo("Illegal symbolic alt allele <DEL>"));
+            assertThat(exception.getMessage(), equalTo("Illegal symbolic alt allele '<DEL>'"));
         }
 
         @Test
         public void throwsExceptionWithBreakendAllele() {
             Exception exception = assertThrows(IllegalArgumentException.class, () -> instance.convert(instance.parseContig("chr1"), "rs123456", 12345, "C", "C[2:321682["));
-            assertThat(exception.getMessage(), equalTo("Illegal symbolic alt allele C[2:321682["));
+            assertThat(exception.getMessage(), equalTo("Illegal symbolic alt allele 'C[2:321682['"));
         }
     }
 
@@ -76,12 +76,12 @@ public class VcfConverterTest {
             // CHR	POS	ID	REF	ALT
             // chr1	12345	.	C	<INS>	6	PASS	SVTYPE=INS;END=12345;SVLEN=200
             GenomicVariant ins = instance.convertSymbolic(instance.parseContig("chr1"), "", 12345, ConfidenceInterval.precise(), 12345, ConfidenceInterval.precise(), "C", "<INS>", 200);
-            assertThat(ins, equalTo(GenomicVariant.of(chr1, "rs123456", Strand.POSITIVE, CoordinateSystem.ONE_BASED, 12345, 12345, "C", "<INS>", 200)));
+            assertThat(ins, equalTo(GenomicVariant.of(chr1, Strand.POSITIVE, CoordinateSystem.ONE_BASED, 12345, 12345, "C", "<INS>", 200)));
         }
 
         @Test
         public void convertSymbolicWithBuilder() {
-            TestGenomicVariant.Builder builder = instance.convertSymbolic(TestGenomicVariant.builder(), instance.parseContig("chr1"), "", 12345, ConfidenceInterval.precise(), 12345, ConfidenceInterval.precise(), "C", "<INS>", 200);
+            GenomicVariant.Builder builder = instance.convertSymbolic(GenomicVariant.builder(), instance.parseContig("chr1"), "", 12345, ConfidenceInterval.precise(), 12345, ConfidenceInterval.precise(), "C", "<INS>", 200);
             GenomicVariant variant = builder.build();
             assertThat(variant.isSymbolic(), equalTo(true));
             assertThat(variant.contig(), equalTo(chr1));
@@ -97,7 +97,7 @@ public class VcfConverterTest {
         @Test
         public void convertSymbolicWithFullyTrimmedRefAlleleBuilder() {
             VcfConverter converter = new VcfConverter(b37, VariantTrimmer.rightShiftingTrimmer(VariantTrimmer.removingCommonBase()));
-            TestGenomicVariant.Builder builder = converter.convertSymbolic(TestGenomicVariant.builder(), instance.parseContig("chr1"), "", 12345, ConfidenceInterval.precise(), 12345, ConfidenceInterval.precise(), "C", "<INS>", 200);
+            GenomicVariant.Builder builder = converter.convertSymbolic(GenomicVariant.builder(), instance.parseContig("chr1"), "", 12345, ConfidenceInterval.precise(), 12345, ConfidenceInterval.precise(), "C", "<INS>", 200);
             GenomicVariant variant = builder.build();
             assertThat(variant.isSymbolic(), equalTo(true));
             assertThat(variant.contig(), equalTo(chr1));
@@ -113,13 +113,18 @@ public class VcfConverterTest {
         @Test
         public void throwsExceptionWithNonSymbolicAllele() {
             Exception exception = assertThrows(IllegalArgumentException.class, () -> instance.convertSymbolic(instance.parseContig("chr1"), "rs123456", 12345, ConfidenceInterval.precise(), 12345, ConfidenceInterval.precise(), "C", "T", 0));
-            assertThat(exception.getMessage(), equalTo("Illegal non-symbolic or breakend alt allele T"));
+            assertThat(exception.getMessage(), equalTo("Illegal non-symbolic alt allele 'T'"));
         }
 
         @Test
-        public void throwsExceptionWithBreakendAllele() {
-            Exception exception = assertThrows(IllegalArgumentException.class, () -> instance.convertSymbolic(instance.parseContig("chr1"), "rs123456", 12345, ConfidenceInterval.precise(), 12345, ConfidenceInterval.precise(), "C", "C[2:321682[", 0));
-            assertThat(exception.getMessage(), equalTo("Illegal non-symbolic or breakend alt allele C[2:321682["));
+        public void convertSymbolicWithBreakendAllele() {
+            GenomicVariant bnd = instance.convertSymbolic(instance.parseContig("chr1"), "bnd_U", 12345, ConfidenceInterval.precise(), 12345, ConfidenceInterval.precise(), "C", "C[2:321682[", 0, "bnd_V", "tra2");
+            assertThat(bnd.isSymbolic(), equalTo(true));
+            assertThat(bnd.isBreakend(), equalTo(true));
+            assertThat(bnd.ref(), equalTo("C"));
+            assertThat(bnd.alt(), equalTo("C[2:321682["));
+            // breakends cannot be flipped from one strand to the other
+            assertSame(bnd, bnd.toOppositeStrand());
         }
     }
 
@@ -151,13 +156,13 @@ public class VcfConverterTest {
         @Test
         public void throwsExceptionWithNonSymbolicAllele() {
             Exception exception = assertThrows(IllegalArgumentException.class, () -> instance.convertBreakend(instance.parseContig("chr1"), "rs123456", 12345, ConfidenceInterval.precise(), "C", "T", ConfidenceInterval.precise(), "", ""));
-            assertThat(exception.getMessage(), equalTo("Illegal non-breakend alt allele T"));
+            assertThat(exception.getMessage(), equalTo("Illegal non-breakend alt allele 'T'"));
         }
 
         @Test
         public void throwsExceptionWithSymbolicAllele() {
             Exception exception = assertThrows(IllegalArgumentException.class, () -> instance.convertBreakend(instance.parseContig("chr1"), "rs123456", 12345, ConfidenceInterval.precise(), "C", "<DEL>", ConfidenceInterval.precise(), "", ""));
-            assertThat(exception.getMessage(), equalTo("Illegal non-breakend alt allele <DEL>"));
+            assertThat(exception.getMessage(), equalTo("Illegal non-breakend alt allele '<DEL>'"));
         }
     }
 }
